continued

+++ include/Key.hpp
@@ -30,7 +30,7 @@ namespace kbd
         static inline constexpr uint8_t maxVelocity{127};               //ffff
         static inline constexpr uint8_t defaultVelocity{64};
 
-        static inline constexpr double slope{(minVelocity - maxVelocity) / maxManipulationTimeMillis};
+        static inline constexpr double slope{static_cast<double>(minVelocity - maxVelocity) / maxManipulationTimeMillis};
 
         uint8_t number;
